+++ src/ByteCodeRelocator.cpp
@@ -5,7 +5,7 @@
 #include <unordered_map>
 #include <iostream>
 
-std::pair<bool, size_t> AddrLocateTable::getOffset(const std::string & symbolName) const {
+std::pair<bool, uint32_t> AddrLocateTable::getOffset(const std::string & symbolName) const {
     auto it = _map.cbegin();
     if ((it = _map.find(symbolName)) != _map.cend()) {
         return std::make_pair(true, it->second);
@@ -13,12 +13,12 @@ std::pair<bool, size_t> AddrLocateTable::getOffset(const std::string & symbolNam
     return std::make_pair(false, 0);
 }
 
-bool AddrLocateTable::setOffset(const std::string & symbolName, size_t offset) {
+bool AddrLocateTable::setOffset(const std::string & symbolName, uint32_t offset) {
     auto it = _map.insert(std::make_pair(symbolName, offset));
     return it.second;
 }
 
-std::pair<bool, size_t>
+std::pair<bool, uint32_t>
 AddrRelocateTable::getOffset(const std::string & symbolName) const {
     auto it = _map.cbegin();
     if ((it = _map.find(symbolName)) != _map.cend()) {
@@ -27,7 +27,7 @@ AddrRelocateTable::getOffset(const std::string & symbolName) const {
     return std::make_pair(false, 0);
 }
 
-void AddrRelocateTable::setOffset(const std::string & symbolName, size_t offset) {
+void AddrRelocateTable::setOffset(const std::string & symbolName, uint32_t offset) {
     auto it = _map.insert(std::make_pair(symbolName, offset));
 }
 
@@ -39,15 +39,15 @@ bool ByteCodeRelocator::relocateStatic(
     // 先重定位静态区块
     for (const auto & e : altStatic._map) {
         const std::string & symbolName = e.first;
-        size_t dwReloc = e.second;
+        uint32_t dwReloc = e.second;
         
         // 遍历重定位表, 找到符号替换其偏移位置的字节码 (DW)
-        //std::pair<bool, size_t> pair;
+        //std::pair<bool, uint32_t> pair;
         auto it = artStatic._map.cbegin();
         while ((it = artStatic._map.find(symbolName)) != artStatic._map.cend()) {
             // 重定位完成后, 删除此条重定位信息
-            size_t offset = it->second;
-            for (size_t i = 0; i < dwReloc; ++i) {
+            uint32_t offset = it->second;
+            for (uint32_t i = 0; i < dwReloc; ++i) {
                 bytesStatic[offset + i] = ((char *)&dwReloc)[i];
             }
             artStatic._map.erase(it);
@@ -73,15 +73,15 @@ bool ByteCodeRelocator::relocateFuncs(
     // 再重定位非静态区块
     for (const auto & e : altFuncs._map) {
         const std::string & symbolName = e.first;
-        size_t dwReloc = e.second;
+        uint32_t dwReloc = e.second;
         
         // 遍历重定位表, 找到符号替换其偏移位置的字节码 (DW)
-        //std::pair<bool, size_t> pair;
+        //std::pair<bool, uint32_t> pair;
         auto it = artFuncs._map.cbegin();
         while ((it = artFuncs._map.find(symbolName)) != artFuncs._map.cend()) {
             // 重定位完成后, 删除此条重定位信息
-            size_t offset = it->second;
-            for (size_t i = 0; i < 4; ++i) {
+            uint32_t offset = it->second;
+            for (uint32_t i = 0; i < 4; ++i) {
                 bytesFuncs[offset + i] = ((char *)&dwReloc)[i];
             }
             artFuncs._map.erase(it);
