+++ src/FakeOPROM.cpp
@@ -7,74 +7,74 @@ FakeOPROM::FakeOPROM(
         _staticByteCodes(staticByteCodes),
         _funcsByteCodes(funcsByteCodes) { }
 
-int8_t FakeOPROM::getStaticB(size_t i) const {
+int8_t FakeOPROM::getStaticB(uint32_t i) const {
     if (i + sizeof(int8_t) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int8_t *>(&_staticByteCodes[i]));
 }
 
-int16_t FakeOPROM::getStaticW(size_t i) const {
+int16_t FakeOPROM::getStaticW(uint32_t i) const {
     if (i + sizeof(int16_t) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int16_t *>(&_staticByteCodes[i]));
 }
 
-int32_t FakeOPROM::getStaticDW(size_t i) const {
+int32_t FakeOPROM::getStaticDW(uint32_t i) const {
     if (i + sizeof(int32_t) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int32_t *>(&_staticByteCodes[i]));
 }
 
-int64_t FakeOPROM::getStaticQW(size_t i) const {
+int64_t FakeOPROM::getStaticQW(uint32_t i) const {
     if (i + sizeof(int64_t) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int64_t *>(&_staticByteCodes[i]));
 }
 
-float FakeOPROM::getStaticFLT(size_t i) const {
+float FakeOPROM::getStaticFLT(uint32_t i) const {
     if (i + sizeof(float) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const float *>(&_staticByteCodes[i]));
 }
 
-double FakeOPROM::getStaticDBL(size_t i) const {
+double FakeOPROM::getStaticDBL(uint32_t i) const {
     if (i + sizeof(double) > _staticByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const double *>(&_staticByteCodes[i]));
 }
 
 
-int8_t FakeOPROM::getFuncsB(size_t i) const {
+int8_t FakeOPROM::getFuncsB(uint32_t i) const {
     if (i + sizeof(int8_t) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int8_t *>(&_funcsByteCodes[i]));
 }
 
-int16_t FakeOPROM::getFuncsW(size_t i) const {
+int16_t FakeOPROM::getFuncsW(uint32_t i) const {
     if (i + sizeof(int16_t) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int16_t *>(&_funcsByteCodes[i]));
 }
 
-int32_t FakeOPROM::getFuncsDW(size_t i) const {
+int32_t FakeOPROM::getFuncsDW(uint32_t i) const {
     if (i + sizeof(int32_t) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int32_t *>(&_funcsByteCodes[i]));
 }
 
-int64_t FakeOPROM::getFuncsQW(size_t i) const {
+int64_t FakeOPROM::getFuncsQW(uint32_t i) const {
     if (i + sizeof(int64_t) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const int64_t *>(&_funcsByteCodes[i]));
 }
 
-float FakeOPROM::getFuncsFLT(size_t i) const {
+float FakeOPROM::getFuncsFLT(uint32_t i) const {
     if (i + sizeof(float) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const float *>(&_funcsByteCodes[i]));
 }
 
-double FakeOPROM::getFuncsDBL(size_t i) const {
+double FakeOPROM::getFuncsDBL(uint32_t i) const {
     if (i + sizeof(double) > _funcsByteCodes.size())
         throw VMException(VMET::E_OPROM_ACCESS_OVERFLOW);
     return *(reinterpret_cast<const double *>(&_funcsByteCodes[i]));
